do not copy config into tests
