+++ src/components/games/bacteria/bot-strategy.spec.ts
@@ -11,6 +11,7 @@ import {
   attackerMove,
   defenderMove
 } from "./bot-strategy";
+import { scatteredStartBoards, adjacentStartBoards } from "./start-boards";
 
 // --- Independent brute-force game solver (small boards only) ---------------
 // attacker moves first; returns true iff the attacker can force a win.
@@ -105,15 +106,6 @@ describe("deficiency equals the true game value (brute force)", () => {
 });
 
 // --- Full-game simulations on the real 9x17 board --------------------------
-const boardWidth = 17;
-const rowCount = 9;
-
-const makeBoard = (starts: number[], goals: number[]): Board => {
-  const bacteria = emptyBoard(rowCount, boardWidth);
-  starts.forEach(c => { bacteria[0][c] = 1; });
-  return { bacteria, goals };
-};
-
 // An adversarial defender: among removals, keep the position hardest for the
 // attacker (prefer staying defender-winning; else slow the attacker down).
 const adversarialDefense = (board: Board): Board => {
@@ -147,37 +139,25 @@ const playAttackerBotVsDefender = (start: Board, maxPlies = 400): "attacker" | "
 };
 
 describe("smart bot plays the 9x17 game optimally", () => {
-  const configs = [
-    { starts: [0, 1, 5, 8, 15], goals: [2, 5, 7, 8] },
-    { starts: [1, 3, 4, 8, 11, 14], goals: [0, 5, 8, 12, 14] },
-    { starts: [3, 7, 9, 12], goals: [4, 5, 9, 10, 11, 13] },
-    { starts: [2, 5, 6, 14, 16], goals: [1, 5, 7, 8, 11] },
-    { starts: [2, 5, 6, 8, 12, 16], goals: [3, 8, 9, 10, 13, 16] },
-    { starts: [2, 5, 12, 16], goals: [4, 6, 7, 11, 12] },
-    { starts: [3, 5, 9, 12], goals: [2, 3, 5, 13, 16] },
-    { starts: [1, 8, 9, 10], goals: [3, 9, 13, 16] },
-    { starts: [0, 4, 5, 11, 12, 13], goals: [4, 7, 10, 12] },
-    { starts: [1, 3, 4, 11, 12, 15], goals: [1, 4, 5, 10, 12, 16] }
-  ];
+  // The real scattered-variant start boards (generateScatteredStartBoard).
+  const boards = scatteredStartBoards();
 
   it("has a mix of attacker- and defender-winning start boards", () => {
-    const values = configs.map(({ starts, goals }) => deficiency(makeBoard(starts, goals)) >= 1);
+    const values = boards.map(board => deficiency(board) >= 1);
     expect(values.some(Boolean)).toBe(true);   // some attacker wins
     expect(values.some(v => !v)).toBe(true);    // some defender wins
   });
 
   it("attacker bot wins from every attacker-winning start against an optimal defender", () => {
-    for (const { starts, goals } of configs) {
-      const board = makeBoard(starts, goals);
+    for (const board of boards) {
       if (deficiency(board) >= 1) {
-        expect(playAttackerBotVsDefender(board)).toBe("attacker");
+        expect(playAttackerBotVsDefender(board), `goals ${board.goals}`).toBe("attacker");
       }
     }
   });
 
   it("defender bot never loses from a defender-winning start against a spreading attacker", () => {
-    for (const { starts, goals } of configs) {
-      const start = makeBoard(starts, goals);
+    for (const start of boards) {
       if (deficiency(start) !== 0) continue;
       let board = cloneDeep(start);
       for (let ply = 0; ply < 400; ply++) {
@@ -206,45 +186,28 @@ describe("smart bot plays the 9x17 game optimally", () => {
 
 // --- Full-game simulations on the real 9x11 adjacent-goals board -----------
 // The width-17 block above only exercises the scattered variant. These are the
-// eight real start boards of the adjacent variant (generateAdjacentStartBoard),
-// which seed bacteria on rows 0-2, to confirm the shared board-driven bot is
-// also optimal at width 11.
-const adjacentConfigs: { row: number; cols: number[]; goals: number[] }[] = [
-  { row: 2, cols: [2, 4, 6, 8], goals: [3, 4, 5, 6, 7] },
-  { row: 2, cols: [3, 5, 7], goals: [3, 4, 5, 6, 7] },
-  { row: 1, cols: [1, 2, 3, 4], goals: [0, 1, 2, 3, 4, 5] },
-  { row: 2, cols: [0, 1, 3], goals: [0, 1, 2, 3, 4, 5] },
-  { row: 0, cols: [2, 3, 7, 9], goals: range(0, 11) },
-  { row: 0, cols: [1, 2, 4, 8], goals: range(0, 11) },
-  { row: 1, cols: [5, 6, 7, 8], goals: [5, 6, 7, 8, 9, 10] },
-  { row: 2, cols: [7, 9, 10], goals: [5, 6, 7, 8, 9, 10] }
-];
-
-const makeAdjacentBoard = ({ row, cols, goals }: { row: number; cols: number[]; goals: number[] }): Board => {
-  const bacteria = emptyBoard(rowCount, 11);
-  cols.forEach(c => { bacteria[row][c] = 1; });
-  return { bacteria, goals };
-};
-
+// real start boards of the adjacent variant (generateAdjacentStartBoard), which
+// seed bacteria on rows 0-2, to confirm the shared board-driven bot is also
+// optimal at width 11.
 describe("smart bot plays the 9x11 adjacent-goals game optimally", () => {
+  const boards = adjacentStartBoards();
+
   it("has a mix of attacker- and defender-winning start boards", () => {
-    const values = adjacentConfigs.map(cfg => deficiency(makeAdjacentBoard(cfg)) >= 1);
+    const values = boards.map(board => deficiency(board) >= 1);
     expect(values.some(Boolean)).toBe(true);
     expect(values.some(v => !v)).toBe(true);
   });
 
   it("attacker bot wins from every attacker-winning start against an optimal defender", () => {
-    for (const cfg of adjacentConfigs) {
-      const board = makeAdjacentBoard(cfg);
+    for (const board of boards) {
       if (deficiency(board) >= 1) {
-        expect(playAttackerBotVsDefender(board), `starts ${cfg.cols} goals ${cfg.goals}`).toBe("attacker");
+        expect(playAttackerBotVsDefender(board), `goals ${board.goals}`).toBe("attacker");
       }
     }
   });
 
   it("defender bot never loses from a defender-winning start against a spreading attacker", () => {
-    for (const cfg of adjacentConfigs) {
-      const start = makeAdjacentBoard(cfg);
+    for (const start of boards) {
       if (deficiency(start) !== 0) continue;
       let board = cloneDeep(start);
       for (let ply = 0; ply < 400; ply++) {
